Rename SCMMaker > SCMManager, add scm client methods
